+++ frontend/src/App.css
@@ -149,230 +149,6 @@
   font-weight: 500;
 }
 
-/* Auth container styles */
-.auth-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  padding: 20px;
-}
-
-.auth-card {
-  background: white;
-  border-radius: 12px;
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
-  padding: 40px;
-  width: 100%;
-  max-width: 400px;
-}
-
-.auth-title {
-  text-align: center;
-  margin-bottom: 30px;
-  color: #333;
-  font-size: 2rem;
-  font-weight: 600;
-}
-
-.btn-primary {
-  background: #1976d2;
-  color: #fff;
-}
-
-.btn-primary:hover {
-  background: #1565c0;
-  transform: translateY(-2px);
-  box-shadow: 0 6px 20px rgba(25, 118, 210, 0.4);
-}
-
-/* Responsive design */
-@media (max-width: 768px) {
-  .navbar {
-    padding: 0 16px;
-    height: 56px;
-  }
-
-  .nav-brand {
-    font-size: 20px;
-  }
-
-  .nav-links {
-    gap: 16px;
-  }
-
-  .nav-link {
-    padding: 6px 12px;
-    font-size: 14px;
-  }
-
-  .homepage {
-    min-height: calc(100vh - 56px);
-    padding: 32px 16px;
-  }
-
-  .homepage h1 {
-    font-size: 36px;
-  }
-
-  .homepage p {
-    font-size: 18px;
-    margin-bottom: 32px;
-  }
-
-  .features {
-    grid-template-columns: 1fr;
-    gap: 24px;
-  }
-
-  .feature-card {
-    padding: 24px;
-  }
-
-  .feature-card h3 {
-    font-size: 20px;
-  }
-}
-
-@media (max-width: 480px) {
-  .homepage h1 {
-    font-size: 28px;
-  }
-
-  .homepage p {
-    font-size: 16px;
-  }
-
-  .feature-card {
-    padding: 20px;
-  }
-
-  .btn {
-    width: 100%;
-    max-width: 200px;
-  }
-}
-
-.auth-form {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-/* Form input styles */
-.form-group {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.form-label {
-  font-weight: 500;
-  color: #555;
-  font-size: 0.9rem;
-}
-
-.required {
-  color: #e74c3c;
-  margin-left: 4px;
-}
-
-.form-input {
-  padding: 12px 16px;
-  border: 2px solid #e1e5e9;
-  border-radius: 8px;
-  font-size: 1rem;
-  transition: all 0.2s ease;
-  background: #fafbfc;
-  color: #333;
-}
-
-.form-input:focus {
-  outline: none;
-  border-color: #667eea;
-  background: white;
-  color: #333;
-  box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1);
-}
-
-.form-input.focused {
-  border-color: #667eea;
-  background: white;
-  color: #333;
-}
-
-.form-input.error {
-  border-color: #e74c3c;
-  background: #fdf2f2;
-  color: #333;
-}
-
-.error-message {
-  color: #e74c3c;
-  font-size: 0.85rem;
-  margin-top: 4px;
-}
-
-.general-error {
-  background: #fdf2f2;
-  border: 1px solid #fecaca;
-  color: #e74c3c;
-  padding: 12px 16px;
-  border-radius: 8px;
-  margin-bottom: 20px;
-  font-size: 0.9rem;
-}
-
-/* Button styles */
-.auth-button {
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  border: none;
-  padding: 14px 24px;
-  border-radius: 8px;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  margin-top: 10px;
-}
-
-.auth-button:hover:not(:disabled) {
-  transform: translateY(-2px);
-  box-shadow: 0 8px 20px rgba(102, 126, 234, 0.3);
-}
-
-.auth-button:disabled {
-  opacity: 0.7;
-  cursor: not-allowed;
-  transform: none;
-}
-
-/* Auth footer */
-.auth-footer {
-  text-align: center;
-  margin-top: 30px;
-  padding-top: 20px;
-  border-top: 1px solid #e1e5e9;
-}
-
-.auth-footer p {
-  color: #666;
-  font-size: 0.9rem;
-}
-
-.auth-link {
-  color: #667eea;
-  text-decoration: none;
-  font-weight: 500;
-  transition: color 0.2s ease;
-}
-
-.auth-link:hover {
-  color: #764ba2;
-  text-decoration: underline;
-}
-
 /* Home page styles */
 .home-container {
   display: flex;
@@ -406,155 +182,69 @@
   font-weight: 600;
 }
 
-.logout-button {
-  background: #e74c3c;
+.account-link {
+  background: #667eea;
   color: white;
-  border: none;
+  text-decoration: none;
   padding: 10px 20px;
   border-radius: 6px;
   font-size: 0.9rem;
   font-weight: 500;
-  cursor: pointer;
   transition: all 0.2s ease;
 }
 
-.logout-button:hover {
-  background: #c0392b;
+.account-link:hover {
+  background: #5a6fd8;
   transform: translateY(-1px);
 }
 
-.user-info {
+.welcome-content {
   margin-bottom: 30px;
 }
 
-.user-info h2 {
+.welcome-message {
+  font-size: 1.1rem;
+  color: #555;
+  margin-bottom: 30px;
+  line-height: 1.6;
+}
+
+.quick-actions h3 {
   color: #333;
   margin-bottom: 20px;
-  font-size: 1.4rem;
+  font-size: 1.2rem;
   font-weight: 600;
 }
 
-.info-group {
+.action-buttons {
   display: flex;
-  align-items: center;
-  margin-bottom: 15px;
-  padding: 15px;
-  background: #f8f9fa;
-  border-radius: 8px;
-}
-
-.info-group label {
-  font-weight: 600;
-  color: #555;
-  min-width: 80px;
-  margin-right: 15px;
-}
-
-.info-group span {
-  color: #333;
-  flex: 1;
-}
-
-.name-display {
-  display: flex;
-  align-items: center;
+  flex-direction: column;
   gap: 15px;
-  flex: 1;
 }
 
-.edit-button {
+.action-button {
   background: #667eea;
   color: white;
   border: none;
-  padding: 6px 12px;
-  border-radius: 4px;
-  font-size: 0.8rem;
+  padding: 12px 20px;
+  border-radius: 8px;
+  font-size: 1rem;
+  font-weight: 500;
   cursor: pointer;
-  transition: background 0.2s ease;
+  transition: all 0.2s ease;
+  text-decoration: none;
+  text-align: center;
 }
 
-.edit-button:hover {
+.action-button:hover:not(.disabled) {
   background: #5a6fd8;
+  transform: translateY(-2px);
 }
 
-.edit-name-container {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  flex: 1;
-}
-
-.edit-name-input {
-  padding: 8px 12px;
-  border: 2px solid #e1e5e9;
-  border-radius: 6px;
-  font-size: 0.9rem;
-  color: #333;
-}
-
-.edit-name-input:focus {
-  outline: none;
-  border-color: #667eea;
-  color: #333;
-}
-
-.edit-buttons {
-  display: flex;
-  gap: 10px;
-}
-
-.save-button {
-  background: #27ae60;
-  color: white;
-  border: none;
-  padding: 6px 12px;
-  border-radius: 4px;
-  font-size: 0.8rem;
-  cursor: pointer;
-  transition: background 0.2s ease;
-}
-
-.save-button:hover:not(:disabled) {
-  background: #229954;
-}
-
-.save-button:disabled {
-  opacity: 0.7;
-  cursor: not-allowed;
-}
-
-.cancel-button {
+.action-button.disabled {
   background: #95a5a6;
-  color: white;
-  border: none;
-  padding: 6px 12px;
-  border-radius: 4px;
-  font-size: 0.8rem;
-  cursor: pointer;
-  transition: background 0.2s ease;
-}
-
-.cancel-button:hover {
-  background: #7f8c8d;
-}
-
-.message {
-  padding: 12px 16px;
-  border-radius: 6px;
-  margin-top: 15px;
-  font-size: 0.9rem;
-}
-
-.message.success {
-  background: #d4edda;
-  border: 1px solid #c3e6cb;
-  color: #155724;
-}
-
-.message.error {
-  background: #f8d7da;
-  border: 1px solid #f5c6cb;
-  color: #721c24;
+  cursor: not-allowed;
+  opacity: 0.7;
 }
 
 .coming-soon {
@@ -578,7 +268,6 @@
 
 /* Responsive design */
 @media (max-width: 480px) {
-  .auth-card,
   .home-card {
     padding: 30px 20px;
   }
@@ -587,16 +276,5 @@
     flex-direction: column;
     gap: 15px;
     text-align: center;
-  }
-
-  .info-group {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 10px;
-  }
-
-  .info-group label {
-    min-width: auto;
-    margin-right: 0;
   }
 }
